core/services/ardupilot_manager/pyproject.toml
[project]
name = "ardupilot_manager"
version = "0.0.1"
description = "AutoPilot service manager."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiocache==0.12.2",
    "appdirs==1.4.4",
    "commonwealth==0.1.0",
    "fastapi==0.105.0",
    "fastapi-versioning==0.9.1",
    "loguru==0.5.3",
    "packaging==20.4",
    "psutil==5.7.2",
    "pydantic==1.10.12",
    "pyelftools==0.30",
    "python-prctl==1.8.1",
    "pyserial==3.5",
    # This dependency needs to be locked since it is used by fastapi
    "python-multipart==0.0.5",
    "smbus2==0.3.0",
    "uvicorn==0.18.0",
    "validators==0.18.2",
]

[tool.uv]
package = false

[tool.uv.sources]
commonwealth = { workspace = true }
